Fix empty response from LLM crash issue
Potential solution for the issue:
https://github.com/t41372/Open-LLM-VTuber/issues/6
Where the MemGPT may give an empty text and crash the program. The empty response from the MemGPT will generate an empty audio file, which would make the max volume in stream_audio to be zero, and therefore would crash the program because we will be divided by zero. Fix it by adding some empty input checks.